Refactored edit mode functions
