+++ main.js
@@ -1,10 +1,12 @@
 import {
-  getSelectedDate,
   renderCalender,
   renderCalenderMonth,
   setDateFact,
 } from "./modules/calender.js";
-import { renderTodoList, loadTodosFromLocalStorage } from "./modules/todo.js";
+import {
+  renderTodoList,
+  loadTodosFromLocalStorage,
+} from "./modules/todos/todo.js";
 import { renderClock } from "./modules/Clock.js";
 
 loadTodosFromLocalStorage();
