+++ modules/todos/todo.js
@@ -1,4 +1,5 @@
-import { renderCalender, getSelectedDate, isSameDate } from "./calender.js";
+import { renderCalender, getSelectedDate, isSameDate } from "../calender.js";
+import { openEditMode } from "./editMode.js";
 
 const todos = [];
 
@@ -131,94 +132,6 @@ function removeTodo(id) {
   renderCalender();
 }
 
-function openEditMode(e, todoDiv) {
-  const todoToEdit = getTodos().find(
-    (todo) => todo.id == e.target.dataset.todoId
-  );
-
-  todoDiv.childNodes.forEach((e) => (e.style.display = "none"));
-
-  const editForm = document.createElement("form");
-  editForm.classList.add("edit-form");
-
-  const titleLabel = document.createElement("label");
-  titleLabel.innerText = "Titel:";
-  const titleEdit = document.createElement("input");
-  titleEdit.type = "text";
-  titleEdit.classList.add("titleEdit");
-  titleEdit.value = todoToEdit.title;
-  titleEdit.required = true;
-
-  // Could not get a pattern to block whitespace so this is a workaround
-  titleEdit.addEventListener("keyup", (e) => {
-    if (e.target.value.trim().length == 0) {
-      e.target.value = "";
-    }
-  });
-
-  editForm.appendChild(titleLabel);
-  editForm.appendChild(titleEdit);
-
-  const descriptionLabel = document.createElement("label");
-  descriptionLabel.innerText = "Beskrivning:";
-  const descriptionEdit = document.createElement("textarea");
-  descriptionEdit.classList.add("descriptionEdit");
-  descriptionEdit.value = todoToEdit.description;
-  descriptionEdit.rows = "3";
-  descriptionEdit.cols = "40";
-  editForm.appendChild(descriptionLabel);
-  editForm.appendChild(descriptionEdit);
-
-  const dateLabel = document.createElement("label");
-  dateLabel.innerText = "Datum:";
-  const dateEdit = document.createElement("input");
-  dateEdit.type = "date";
-  const dateValue = todoToEdit.date.toISOString().split("T")[0];
-  dateEdit.value = dateValue;
-  dateEdit.required = true;
-  dateEdit.classList.add("dateEdit");
-
-  const btnDiv = document.createElement("div");
-  btnDiv.classList.add("editmodeBtnDiv");
-
-  const returnBtn = document.createElement("button");
-  returnBtn.innerHTML = "Återgå";
-  returnBtn.classList.add("returnBtn");
-  returnBtn.addEventListener("click", () => {
-    renderTodoList(getSelectedDate());
-  });
-
-  const saveBtn = document.createElement("button");
-  saveBtn.classList.add("saveBtn");
-  saveBtn.innerHTML = "Spara";
-
-  btnDiv.appendChild(returnBtn);
-  btnDiv.appendChild(saveBtn);
-
-  const todoId = e.target.dataset.todoId;
-
-  editForm.addEventListener("submit", (e) =>
-    saveEdit(dateEdit.value, titleEdit.value, descriptionEdit.value, todoId, e)
-  );
-
-  editForm.appendChild(dateLabel);
-  editForm.appendChild(dateEdit);
-  editForm.appendChild(btnDiv);
-  todoDiv.appendChild(editForm);
-}
-
-function saveEdit(date, title, description, id, event) {
-  event.preventDefault();
-  const todo = todos.find((todo) => todo.id == id);
-  todo.date = new Date(date);
-  todo.title = title;
-  todo.description = description;
-
-  saveTodosToLocalStorage();
-  renderTodoList(getSelectedDate());
-  renderCalender();
-}
-
 function filterTodosByDate() {
   const filteredTodoArray = todos.filter((todo) => {
     const currentTime = new Date();
@@ -264,4 +177,9 @@ function loadTodosFromLocalStorage() {
   }
 }
 
-export { renderTodoList, getTodos, loadTodosFromLocalStorage };
+export {
+  renderTodoList,
+  getTodos,
+  loadTodosFromLocalStorage,
+  saveTodosToLocalStorage,
+};
